refactor(dashboard): 将 Memory 与 Interface Dashboard 合并至 Home Dashboard 统一入口;
- 新建 MemoryOverviewSection / InterfaceOverviewSection 组件，
  将原 Memory Dashboard（8 指标卡片 + Retrieval Metrics）和
  Interface Dashboard（StatCards + QuickLinks）内容以 Section 形式
  嵌入 Home Dashboard 页面底部，三类数据并行独立获取，零耦合；
- /memory 与 /interface 路由改为 router.replace("/dashboard") 重定向，
  保留书签兼容性，Memory / Interface 各子页面不受影响；
- MemoryNav / InterfaceNav 的 Dashboard tab href 统一指向 /dashboard；
- 主导航 config 中 Memory → /memory/timeline、Interface → /interface/subagents，
  避免指向重定向壳页面；
- 同步更新 admin roles 页面引用及 admin 回退重定向路径；
- 同步更新单元测试与 e2e 测试断言。
