+++ apps/negentropy-ui/app/interface/task-models/page.tsx
@@ -49,7 +49,7 @@ export default function TaskModelsPage() {
   useEffect(() => {
     if (status === "loading") return;
     if (!user?.roles?.includes("admin")) {
-      router.replace("/interface");
+      router.replace("/interface/subagents");
     }
   }, [user, status, router]);
