+++ apps/negentropy-ui/components/ui/InterfaceNav.tsx
@@ -10,7 +10,7 @@ import { useAuth } from "@/components/providers/AuthProvider";
 type NavItem = { href: string; label: string; adminOnly?: boolean };
 
 const NAV_ITEMS: NavItem[] = [
-  { href: "/interface", label: "Dashboard" },
+  { href: "/dashboard", label: "Dashboard" },
   { href: "/interface/models", label: "Models", adminOnly: true },
   { href: "/interface/task-models", label: "Task Models", adminOnly: true },
   { href: "/interface/subagents", label: "SubAgents" },
@@ -38,7 +38,9 @@ export function InterfaceNav({
   }, [title, setNavigationInfo]);
 
   const isActive = (href: string) =>
-    href === "/interface" ? pathname === "/interface" : pathname.startsWith(href);
+    href === "/dashboard"
+      ? pathname === "/dashboard"
+      : pathname.startsWith(href);
 
   const visibleItems = NAV_ITEMS.filter((item) => !item.adminOnly || isAdmin);
