+++ apps/negentropy-ui/components/ui/MemoryNav.tsx
@@ -6,7 +6,7 @@ import { usePathname } from "next/navigation";
 import { useNavigation } from "@/components/providers/NavigationProvider";
 
 const NAV_ITEMS = [
-  { href: "/memory", label: "Dashboard" },
+  { href: "/dashboard", label: "Dashboard" },
   { href: "/memory/timeline", label: "Timeline" },
   { href: "/memory/facts", label: "Facts" },
   { href: "/memory/audit", label: "Audit" },
@@ -32,7 +32,9 @@ export function MemoryNav({
   }, [title, setNavigationInfo]);
 
   const isActive = (href: string) =>
-    href === "/memory" ? pathname === "/memory" : pathname.startsWith(href);
+    href === "/dashboard"
+      ? pathname === "/dashboard"
+      : pathname.startsWith(href);
 
   return (
     <div className="border-b border-border bg-card px-6 py-1">
