+++ apps/negentropy-ui/tests/unit/ui/InterfaceNav.test.tsx
@@ -38,7 +38,7 @@ describe("InterfaceNav", () => {
 
     expect(screen.getByRole("link", { name: "Dashboard" })).toHaveAttribute(
       "href",
-      "/interface",
+      "/dashboard",
     );
   });
